remove common preference class
